node-main/deps/v8/test/webkit/fast/js/Promise-static-race-expected.txt

Test Promise.race

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS result is undefined
PASS Promise.race() is rejected.
PASS Promise.race({}) is rejected.
PASS Promise.race([p4, p1, p6]) is fulfilled.
PASS result is "p1"
PASS Promise.race([p4, p6, p1]) is rejected.
PASS result is "p6"
PASS Promise.race([p9]) is fulfilled.
PASS result is "p2"
PASS Promise.race([p4,,]) is fulfilled.
PASS result is undefined
PASS Promise.race([p4,42]) is fulfilled.
PASS result is 42
PASS successfullyParsed is true

TEST COMPLETE
